Add support for geojson serializer
