+++ api/db.py
@@ -1,35 +1,10 @@
-import json
-
 from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.ext.declarative import DeclarativeMeta
 from sqlalchemy.orm import scoped_session
 from sqlalchemy.orm import sessionmaker
 
 
 Base = declarative_base()
-
-
-class AlchemyEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj.__class__, DeclarativeMeta):
-            # A SQLAlchemy class
-            fields = {}
-            for field in [
-                x for x in dir(obj) if not x.startswith("_") and x != "metadata"
-            ]:
-                data = obj.__getattribute__(field)
-                try:
-                    json.dumps(
-                        data
-                    )  # this will fail on non-encodable values, like other classes
-                    fields[field] = data
-                except TypeError:
-                    fields[field] = None
-            # A json-encodable dict
-            return fields
-
-        return json.JSONEncoder.default(self, obj)
 
 
 def get_db(db_uri):
